Assorted cleanup, warning fixes

+++ parts/components/w25x20cl.cpp
@@ -21,12 +21,13 @@
 	along with MK3SIM.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include <stdlib.h>
-#include <stdio.h>
-#include <string.h>
-#include <unistd.h>
-#include <fcntl.h>
 #include "w25x20cl.h"
+#include <fcntl.h>      // for open, O_CREAT, O_RDWR, SEEK_SET
+#include <stdio.h>      // for printf, perror, fprintf, stderr, size_t
+#include <stdlib.h>     // for exit, free, malloc
+#include <string.h>     // for memset, memcpy, strncpy
+#include <sys/types.h>  // for uint
+#include <unistd.h>     // for close, ftruncate, lseek, read, write, ssize_t
 
 //#define TRACE(_w) _w
 #ifndef TRACE
@@ -78,7 +79,7 @@ uint8_t w25x20cl::OnSPIIn(struct avr_irq_t * irq, uint32_t value)
 			if (m_rxCnt >= sizeof(m_cmdIn))
 			{
 				printf("w25x20cl_t: error: command too long: ");
-				for (int i = 0; i < sizeof(m_cmdIn); i++)
+				for (size_t i = 0; i < sizeof(m_cmdIn); i++)
 				{
 					printf("%02x, ", m_cmdIn[i]);
 				}
@@ -87,8 +88,8 @@ uint8_t w25x20cl::OnSPIIn(struct avr_irq_t * irq, uint32_t value)
 			}
 			m_cmdIn[m_rxCnt] = value;
 			m_rxCnt++;
-			
-			
+
+
 			// Check for a loaded instruction in the cmdIn buffer
 			m_command = m_cmdIn[0];
 			switch(m_command)
@@ -111,7 +112,7 @@ uint8_t w25x20cl::OnSPIIn(struct avr_irq_t * irq, uint32_t value)
 						m_state = STATE_RUNNING;
 					}
 				} break;
-				
+
 				case _CMD_PAGE_PROGRAM:
 				{
 					if (m_rxCnt == 4)
@@ -127,7 +128,7 @@ uint8_t w25x20cl::OnSPIIn(struct avr_irq_t * irq, uint32_t value)
 						m_state = STATE_RUNNING;
 					}
 				} break;
-				
+
 				case _CMD_ENABLE_WR:
 				case _CMD_DISABLE_WR:
 				case _CMD_RD_STATUS_REG:
@@ -136,7 +137,7 @@ uint8_t w25x20cl::OnSPIIn(struct avr_irq_t * irq, uint32_t value)
 				{
 					m_state = STATE_RUNNING;
 				} break;
-				
+
 				case _CMD_RD_UID:
 				{
 					if (m_rxCnt == 5)
@@ -145,7 +146,7 @@ uint8_t w25x20cl::OnSPIIn(struct avr_irq_t * irq, uint32_t value)
 						m_state = STATE_RUNNING;
 					}
 				} break;
-				
+
 				default:
 				{
 				printf("w25x20cl_t: error: unknown command: ");
@@ -156,7 +157,7 @@ uint8_t w25x20cl::OnSPIIn(struct avr_irq_t * irq, uint32_t value)
 				printf("\n");
 				} break;
 			}
-			
+
 		} break;
 		case STATE_RUNNING:
 		{
@@ -199,7 +200,7 @@ uint8_t w25x20cl::OnSPIIn(struct avr_irq_t * irq, uint32_t value)
 			}
 			break;
 		}
-		
+
 		case STATE_IDLE:
 		default:
 		{
@@ -330,4 +331,4 @@ void w25x20cl::Save()
 		fprintf(stderr, "unable to write xflash memory\n");
 		perror(m_filepath);
 	}
-} 
+}
